DumprX: Randomize Branch Names
(1). This is needed in order to dump custom-roms, as multiple builds of the same device might use the same fingerprint, thus not pushing the firmware, or git notifying to use "-f" flag, which is not what we want.
(2). We aren't hosting our dumps on a custom server, so storage isn't much of a headache as of now.

Change-Id: If4fe3d9973d3e299be662b61afea1b32473e71e3
Signed-off-by: Renayura <renayura@proton.me>